fix(admin, bot): use float64 and prevent finding similarity between diff shapes

+++ bot/ai_utils.py
@@ -5,5 +5,5 @@ def embedding_to_bytes(embedding: np.ndarray) -> bytes:
     return embedding.tobytes()
 
 
-def embedding_from_bytes(embedding: bytes) -> np.ndarray:
-    return np.frombuffer(embedding, dtype=np.float32)
+def embedding_from_bytes(embedding: bytes, dtype=np.float64) -> np.ndarray:
+    return np.frombuffer(embedding, dtype=dtype)
